% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{filter_freeze_directory}
\alias{filter_freeze_directory}
\title{Filter the freeze directory}
\usage{
filter_freeze_directory(temporary_directory, project_name, language_code)
}
\arguments{
\item{temporary_directory}{Temporary directory where the project is}

\item{project_name}{Name of the project directory}

\item{language_code}{The Language code for the current rendering}
}
\description{
This function removes from freeze directory the files
that do not match the language code to avoid using the
wrong cache values for code computations.
}
