% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-yaml.R
\name{read_yaml}
\alias{read_yaml}
\title{Alias for yaml::read_yaml with a custom handler}
\usage{
read_yaml(path)
}
\arguments{
\item{path}{Path to yaml file}
}
\value{
A list
}
\description{
Alias for yaml::read_yaml with a custom handler
}
\examples{
\dontrun{
read_yaml(path)
}
}
