% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_further_languages}
\alias{register_further_languages}
\title{Register further languages in Quarto config}
\usage{
register_further_languages(further_languages, project_path = ".")
}
\arguments{
\item{further_languages}{Languages to be registered (character vector)}

\item{project_path}{Path where the book/website/presentation source is
located.}
}
\value{
Nothing, called for its side-effect of editing the Quarto configuration file.
}
\description{
Register further languages in Quarto config
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_dir <- withr::local_tempdir()
  quarto_multilingual_book(
    parent_dir = parent_dir,
    project_dir = "blop",
    further_languages = c("es", "fr"),
    main_language = "en"
  )
register_further_languages(
  "pt",
  project_path = file.path(parent_dir, "blop")
)
# have a look at the configuration
file.edit(file.path(parent_dir, "blop", "_quarto.yml"))
\dontshow{\}) # examplesIf}
}
