% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-dynamic.R
\name{render_dynamic_book}
\alias{render_dynamic_book}
\alias{render_dynamic_website}
\title{Render a Quarto multilingual project}
\usage{
render_dynamic_book(project_path = ".", site_url = NULL)

render_dynamic_website(project_path = ".", site_url = NULL)
}
\arguments{
\item{project_path}{Path where the book/website source is located}

\item{site_url}{Base URL of the book/website.}
}
\description{
Render a Quarto multilingual project
}
\details{
babelquarto expects a book/website folder with
mixed qmd/Rmd present when chapter in language versions are different,
\itemize{
\item once with only \code{.qmd} as extension for the main language,
\item once with \code{.es.qmd} (using the language code) for each other language.
}

You also need to register the language in the configuration file,
see \code{\link[=register_main_language]{register_main_language()}}
and \code{\link[=register_further_languages]{register_further_languages()}}:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{babelquarto:
  mainlanguage: 'en'
  languages: ['es', 'fr']
}\if{html}{\out{</div>}}
}
\examples{
directory <- withr::local_tempdir()
quarto_multilingual_book(parent_dir = directory, project_dir = "blop")
render_dynamic_book(file.path(directory, "blop"))
\dontrun{
if (require("servr") && rlang::is_interactive()) {
  servr::httw(file.path(directory, "blop", "_book"))
}
}

}
