% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-quarto.R
\name{quarto_multilingual_book}
\alias{quarto_multilingual_book}
\title{Create a starter/example quarto multilingual book}
\usage{
quarto_multilingual_book(
  parent_dir,
  book_dir,
  main_language = "en",
  further_languages = c("es", "fr"),
  register_languages = TRUE,
  site_url = "https://example.com"
)
}
\arguments{
\item{parent_dir}{Folder where to create the book folder.}

\item{book_dir}{Book folder name.}

\item{main_language}{Code for main languages.}

\item{further_languages}{Codes for not main languages.}

\item{register_languages}{Whether to register languages (logical).}

\item{site_url}{Site URL for the book/site-url part of the Quarto configuration.}
}
\value{
Nothing, creates the book folder.
}
\description{
Create a starter/example quarto multilingual book
}
