% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-quarto.R
\name{quarto_multilingual_book}
\alias{quarto_multilingual_book}
\alias{quarto_multilingual_website}
\title{Create a starter/example quarto multilingual book or website}
\usage{
quarto_multilingual_book(
  parent_dir,
  project_dir,
  main_language = "en",
  further_languages = c("es", "fr"),
  register_languages = TRUE,
  site_url = "https://example.com"
)

quarto_multilingual_website(
  parent_dir,
  project_dir,
  main_language = "en",
  further_languages = c("es", "fr"),
  register_languages = TRUE,
  site_url = "https://example.com"
)
}
\arguments{
\item{parent_dir}{Folder where to create the project folder.}

\item{project_dir}{Project (book, webiste) folder name.}

\item{main_language}{Code for main languages.}

\item{further_languages}{Codes for not main languages.}

\item{register_languages}{Whether to register languages (logical).}

\item{site_url}{Site URL for the book/site-url
or website/site-url part of the Quarto configuration.}
}
\value{
The path to the created project.
}
\description{
Create a starter/example quarto multilingual book or website
}
