% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-quarto.R
\name{quarto_multilingual_book}
\alias{quarto_multilingual_book}
\title{Create a starter/example quarto multilingual book}
\usage{
quarto_multilingual_book(
  parent_dir,
  book_dir,
  further_languages = c("es", "fr"),
  main_language = "en"
)
}
\arguments{
\item{parent_dir}{Folder where to create the book folder.}

\item{book_dir}{Book folder name.}

\item{further_languages}{Codes for not main languages.}

\item{main_language}{Code for main languages.}
}
\value{
Nothing, creates the book folder.
}
\description{
Create a starter/example quarto multilingual book
}
