% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_further_languages}
\alias{register_further_languages}
\title{Register further languages in Quarto config}
\usage{
register_further_languages(further_languages, project_path = ".")
}
\arguments{
\item{further_languages}{Languages to be registered (character vector)}

\item{project_path}{Path where the book/website source is located}
}
\value{
Nothing
}
\description{
Register further languages in Quarto config
}
