% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_main_language}
\alias{register_main_language}
\title{Register main language in Quarto config}
\usage{
register_main_language(main_language = "en", project_path = ".")
}
\arguments{
\item{main_language}{Main language code (character, like \code{"en"})}

\item{project_path}{Path where the book/website source is located}
}
\value{
Nothing
}
\description{
Register main language in Quarto config
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_dir <- withr::local_tempdir()
quarto_multilingual_book(
    parent_dir = parent_dir,
    project_dir = "blop",
    further_languages = c("es", "fr"),
    main_language = "en",
    register_languages = FALSE
  )
book_path <- file.path(parent_dir, "blop")
register_main_language("en", book_path)
# have a look at the config
file.edit(file.path(parent_dir, "blop", "_quarto.yml"))
\dontshow{\}) # examplesIf}
}
