% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_main_language}
\alias{register_main_language}
\title{Register main language in Quarto config}
\usage{
register_main_language(main_language = "en", book_path = ".")
}
\arguments{
\item{main_language}{Main language code (character, like \code{"en"})}

\item{book_path}{Path where the book source is located}
}
\value{
Nothing
}
\description{
Register main language in Quarto config
}
