% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_book}
\alias{render_book}
\alias{render_website}
\title{Render a Quarto multilingual project}
\usage{
render_book(
  project_path = ".",
  site_url = NULL,
  profile = NULL,
  preview = rlang::is_interactive()
)

render_website(
  project_path = ".",
  site_url = NULL,
  profile = NULL,
  preview = rlang::is_interactive()
)
}
\arguments{
\item{project_path}{Path where the book/website source is located}

\item{site_url}{Override the base URL of the book/website.
If \code{NULL}, in interactive sessions it will be set to "" to allow
previewing the whole project with \code{servr::httw()}.}

\item{profile}{Quarto profile(s) to use.}

\item{preview}{Logical indicating whether to preview the project using
\code{servr::httw()}.`}
}
\value{
Nothing, called for its side-effect of rendering a project.
}
\description{
Render a Quarto multilingual project
}
\details{
babelquarto expects a book/website folder with
each qmd/Rmd present in as many languages as needed,
with the same basename but,
\itemize{
\item once with only \code{.qmd} as extension for the main language,
\item once with \code{.es.qmd} (using the language code) for each other language.
}

You also need to register the language in the configuration file,
see \code{\link[=register_main_language]{register_main_language()}}
and \code{\link[=register_further_languages]{register_further_languages()}}:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{babelquarto:
  mainlanguage: 'en'
  languages: ['es', 'fr']
}\if{html}{\out{</div>}}
}
\examples{
directory <- withr::local_tempdir()
quarto_multilingual_book(parent_dir = directory, project_dir = "blop")
render_book(file.path(directory, "blop"))
\dontrun{
if (require("servr") && rlang::is_interactive()) {
  servr::httw(file.path(directory, "blop", "_book"))
}
}

}
