% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_book}
\alias{render_book}
\title{Render a Quarto multilingual book}
\usage{
render_book(book_path = ".", site_url = NULL)
}
\arguments{
\item{book_path}{Path where the book source is located}

\item{site_url}{Base URL of the book.}
}
\description{
Render a Quarto multilingual book
}
\details{
babelquarto expects a book folder with
each qmd/Rmd present in as many languages as needed,
with the same basename but,
\itemize{
\item once with only \code{.qmd} as extension for the main language,
\item once with \code{.es.qmd} (using the language code) for each other language.
}

You also need to register the language in the configuration file:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{babelquarto:
  mainlanguage: 'en'
  languages: ['es', 'fr']
}\if{html}{\out{</div>}}
}
\examples{
directory <- withr::local_tempdir()
quarto_multilingual_book(parent_dir = directory, book_dir = "blop")
render_book(file.path(directory, "blop"))
\dontrun{
if (require("servr") && rlang::is_interactive()) {
  servr::httw(file.path(directory, "blop", "_book"))
}
}

}
